Pull out the GET components from S3ObjectStore as a separate class
